feat: move mail into standalone layout
