--- src/app/(main)/dashboard/mail/_components/mail-layout-config.ts
@@ -1,10 +1,0 @@
-export const MAIL_LAYOUT_COOKIE = "mail_panel_layout";
-export const MAIL_COLLAPSED_COOKIE = "mail_sidebar_collapsed";
-
-export const MAIL_NAV_PANEL_ID = "mail-navigation-panel";
-export const MAIL_LIST_PANEL_ID = "mail-message-list-panel";
-export const MAIL_DETAIL_PANEL_ID = "mail-message-detail-panel";
-
-export const DEFAULT_MAIL_LAYOUT = [22, 30, 48] as const;
-export const DEFAULT_MAIL_COLLAPSED = false;
-export const MAIL_SIDEBAR_COLLAPSED_SIZE = "50px";
