--- src/app/(main)/dashboard/mail/_components/mail.tsx
@@ -1,130 +1,0 @@
-"use client";
-
-import * as React from "react";
-
-import { Drawer, DrawerContent, DrawerTitle } from "@/components/ui/drawer";
-import { ResizableHandle, ResizablePanel, ResizablePanelGroup } from "@/components/ui/resizable";
-import { useIsMobile } from "@/hooks/use-mobile";
-import { setClientCookie } from "@/lib/cookie.client";
-import { cn } from "@/lib/utils";
-
-import type { Mail } from "./data";
-import { MailInbox } from "./mail-inbox";
-import {
-  DEFAULT_MAIL_COLLAPSED,
-  DEFAULT_MAIL_LAYOUT,
-  MAIL_COLLAPSED_COOKIE,
-  MAIL_DETAIL_PANEL_ID,
-  MAIL_LAYOUT_COOKIE,
-  MAIL_LIST_PANEL_ID,
-  MAIL_NAV_PANEL_ID,
-  MAIL_SIDEBAR_COLLAPSED_SIZE,
-} from "./mail-layout-config";
-import { MailSidebar } from "./mail-sidebar";
-import { MailView } from "./mail-view";
-import { useMail } from "./use-mail";
-
-interface MailProps {
-  accounts: {
-    id: number;
-    label: string;
-    email: string;
-  }[];
-  mails: Mail[];
-  defaultLayout: number[] | undefined;
-  defaultCollapsed?: boolean;
-}
-
-export function MailComponent({
-  accounts,
-  mails,
-  defaultLayout = [...DEFAULT_MAIL_LAYOUT],
-  defaultCollapsed = DEFAULT_MAIL_COLLAPSED,
-}: MailProps) {
-  const isMobile = useIsMobile();
-
-  return isMobile ? (
-    <MailMobileLayout accounts={accounts} mails={mails} />
-  ) : (
-    <MailDesktopLayout
-      accounts={accounts}
-      mails={mails}
-      defaultLayout={defaultLayout}
-      defaultCollapsed={defaultCollapsed}
-    />
-  );
-}
-
-function MailMobileLayout({ accounts, mails }: Pick<MailProps, "accounts" | "mails">) {
-  const [mail] = useMail();
-  const [isSidebarOpen, setIsSidebarOpen] = React.useState(false);
-  const [isMailOpen, setIsMailOpen] = React.useState(false);
-  const selectedMail = mails.find((item) => item.id === mail.selected) || null;
-
-  return (
-    <>
-      <MailInbox mails={mails} onOpenSidebar={() => setIsSidebarOpen(true)} onSelectMail={() => setIsMailOpen(true)} />
-
-      <Drawer direction="left" open={isSidebarOpen} onOpenChange={setIsSidebarOpen}>
-        <DrawerContent>
-          <DrawerTitle className="sr-only">Mail navigation</DrawerTitle>
-          <MailSidebar isCollapsed={false} accounts={accounts} />
-        </DrawerContent>
-      </Drawer>
-
-      <Drawer open={isMailOpen} onOpenChange={setIsMailOpen}>
-        <DrawerContent className="data-[vaul-drawer-direction=bottom]:h-svh data-[vaul-drawer-direction=bottom]:max-h-svh">
-          <DrawerTitle className="sr-only">Mail message</DrawerTitle>
-          <MailView mail={selectedMail} onClose={() => setIsMailOpen(false)} />
-        </DrawerContent>
-      </Drawer>
-    </>
-  );
-}
-
-function MailDesktopLayout({
-  accounts,
-  mails,
-  defaultLayout = [...DEFAULT_MAIL_LAYOUT],
-  defaultCollapsed = DEFAULT_MAIL_COLLAPSED,
-}: MailProps) {
-  const [isCollapsed, setIsCollapsed] = React.useState(defaultCollapsed);
-  const [mail] = useMail();
-  const collapsedSize = Number.parseInt(MAIL_SIDEBAR_COLLAPSED_SIZE, 10);
-
-  return (
-    <ResizablePanelGroup
-      orientation="horizontal"
-      onLayoutChanged={(layout) => {
-        const sizes = [layout[MAIL_NAV_PANEL_ID], layout[MAIL_LIST_PANEL_ID], layout[MAIL_DETAIL_PANEL_ID]];
-        setClientCookie(MAIL_LAYOUT_COOKIE, JSON.stringify(sizes));
-      }}
-      className="h-full"
-    >
-      <ResizablePanel
-        id={MAIL_NAV_PANEL_ID}
-        defaultSize={defaultCollapsed ? MAIL_SIDEBAR_COLLAPSED_SIZE : `${defaultLayout[0]}%`}
-        collapsedSize={MAIL_SIDEBAR_COLLAPSED_SIZE}
-        collapsible={true}
-        minSize="15%"
-        maxSize="22%"
-        onResize={(size) => {
-          const shouldCollapse = size.inPixels <= collapsedSize + 1;
-          setIsCollapsed(shouldCollapse);
-          setClientCookie(MAIL_COLLAPSED_COOKIE, JSON.stringify(shouldCollapse));
-        }}
-        className={cn(isCollapsed && "min-w-12.5 transition-all duration-300 ease-in-out")}
-      >
-        <MailSidebar isCollapsed={isCollapsed} accounts={accounts} />
-      </ResizablePanel>
-      <ResizableHandle withHandle />
-      <ResizablePanel id={MAIL_LIST_PANEL_ID} defaultSize={`${defaultLayout[1]}%`} minSize="30%" className="min-h-0">
-        <MailInbox mails={mails} />
-      </ResizablePanel>
-      <ResizableHandle withHandle />
-      <ResizablePanel id={MAIL_DETAIL_PANEL_ID} defaultSize={`${defaultLayout[2]}%`} minSize="30%" className="min-h-0">
-        <MailView mail={mails.find((item) => item.id === mail.selected) || null} />
-      </ResizablePanel>
-    </ResizablePanelGroup>
-  );
-}
