+++ src/app/(main)/mail/_components/mail-layout-config.ts
@@ -1,0 +1,6 @@
+export const MAIL_LAYOUT_COOKIE = "mail_panel_layout";
+
+export const MAIL_LIST_PANEL_ID = "mail-message-list-panel";
+export const MAIL_DETAIL_PANEL_ID = "mail-message-detail-panel";
+
+export const DEFAULT_MAIL_LAYOUT = [38, 62] as const;
